+++ test/networkInterface.ts
@@ -2,6 +2,8 @@ import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 
 import { assign } from 'lodash';
+import isequal = require('lodash.isequal');
+import * as fetchMock from 'fetch-mock';
 
 // make it easy to assert with promises
 chai.use(chaiAsPromised);
@@ -10,9 +12,8 @@ const { assert, expect } = chai;
 
 import {
   createNetworkInterface,
-  addQueryMerging,
-  NetworkInterface,
-  Request,
+//  NetworkInterface,
+//  Request,
 } from '../src/transport/networkInterface';
 
 import {
@@ -27,41 +28,110 @@ import gql from 'graphql-tag';
 
 import { print } from 'graphql-tag/printer';
 
-import { GraphQLResult } from 'graphql';
+// import { GraphQLResult } from 'graphql';
 
 describe('network interface', () => {
+  const swapiUrl = 'http://graphql-swapi.test/';
+  const missingUrl = 'http://does-not-exist.test/';
+
+  const simpleQueryWithNoVars = gql`
+    query people {
+      allPeople(first: 1) {
+        people {
+          name
+        }
+      }
+    }
+  `;
+
+  const simpleQueryWithVar = gql`
+    query people($personNum: Int!) {
+      allPeople(first: $personNum) {
+        people {
+          name
+        }
+      }
+    }
+  `;
+
+  const simpleResult = {
+    data: {
+      allPeople: {
+        people: [
+          {
+            name: 'Luke Skywalker',
+          },
+        ],
+      },
+    },
+  };
+
+  const complexQueryWithTwoVars = gql`
+    query people($personNum: Int!, $filmNum: Int!) {
+      allPeople(first: $personNum) {
+        people {
+          name
+          filmConnection(first: $filmNum) {
+            edges {
+              node {
+                id
+              }
+            }
+          }
+        }
+      }
+    }
+  `;
+
+  const complexResult = {
+    data: {
+      allPeople: {
+        people: [
+          {
+            name: 'Luke Skywalker',
+            filmConnection: {
+              edges: [
+                {
+                  node: {
+                    id: 'ZmlsbXM6MQ==',
+                  },
+                },
+              ],
+            },
+          },
+        ],
+      },
+    },
+  };
+
+  // We mock the network interface to return the results that the SWAPI would.
   before(() => {
-    this.realFetch = (<any>global)['fetch'];
+    // We won't be too careful about counting calls or closely checking
+    // parameters, but just do the basic stuff to ensure the request looks right
+    fetchMock.post(swapiUrl, (url, opts) => {
+      const { query, variables } = JSON.parse((opts as RequestInit).body.toString());
 
-    (<any>global)['fetch'] = ((url: string, opts: any) => {
-      this.lastFetchOpts = opts;
-      if (url === 'http://does-not-exist.test/') {
-        return Promise.reject('Network error');
+      if (query === print(simpleQueryWithNoVars)) {
+        return simpleResult;
       }
 
-      if (url === 'http://graphql-swapi.test/') {
-        url = 'http://graphql-swapi.parseapp.com/';
+      if (query === print(simpleQueryWithVar)
+          && isequal(variables, { personNum: 1 })) {
+        return simpleResult;
       }
 
-      // XXX swapi graphql NPM package is broken now
-      // else if (url === 'http://graphql-swapi.test/') {
-      //   return new Promise((resolve, reject) => {
-      //     const request = JSON.parse(opts.body);
-      //     graphql(swapiSchema, request.query, undefined, request.variables).then(result => {
-      //       const response = new (<any>global)['Response'](JSON.stringify(result));
-      //       resolve(response);
-      //     }).catch(error => {
-      //       reject(error);
-      //     });
-      //   });
-      // }
+      if (query === print(complexQueryWithTwoVars)
+          && isequal(variables, { personNum: 1, filmNum: 1 })) {
+        return complexResult;
+      }
 
-      return this.realFetch(url, opts);
+      throw new Error('Invalid Query');
     });
+    fetchMock.post(missingUrl, Promise.reject('Network error'));
   });
 
   after(() => {
-    (<any>global)['fetch'] = this.realFetch;
+    fetchMock.restore();
   });
 
   describe('creating a network interface', () => {
@@ -72,7 +142,7 @@ describe('network interface', () => {
     });
 
     it('should create an instance with a given uri', () => {
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
       assert.equal(networkInterface._uri, '/graphql');
     });
 
@@ -82,7 +152,7 @@ describe('network interface', () => {
         credentials: 'include',
       };
 
-      const networkInterface = createNetworkInterface('/graphql', customOpts);
+      const networkInterface = createNetworkInterface({ uri: '/graphql', opts: customOpts });
 
       assert.deepEqual(networkInterface._opts, assign({}, customOpts));
     });
@@ -94,7 +164,7 @@ describe('network interface', () => {
       };
       const originalOpts = assign({}, customOpts);
 
-      const networkInterface = createNetworkInterface('/graphql', customOpts);
+      const networkInterface = createNetworkInterface({ uri: '/graphql', opts: customOpts });
 
       delete customOpts.headers;
 
@@ -105,7 +175,7 @@ describe('network interface', () => {
   describe('middleware', () => {
     it('should throw an error if you pass something bad', () => {
       const malWare: any = {};
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
 
       try {
         networkInterface.use([malWare]);
@@ -122,7 +192,7 @@ describe('network interface', () => {
     it('should take a middleware and assign it', () => {
       const testWare = TestWare();
 
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
       networkInterface.use([testWare]);
 
       assert.equal(networkInterface._middlewares[0], testWare);
@@ -132,7 +202,7 @@ describe('network interface', () => {
       const testWare1 = TestWare();
       const testWare2 = TestWare();
 
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
       networkInterface.use([testWare1, testWare2]);
 
       assert.deepEqual(networkInterface._middlewares, [testWare1, testWare2]);
@@ -143,36 +213,18 @@ describe('network interface', () => {
         { key: 'personNum', val: 1 },
       ]);
 
-      const swapi = createNetworkInterface('http://graphql-swapi.test/');
+      const swapi = createNetworkInterface({ uri: swapiUrl });
       swapi.use([testWare1]);
       // this is a stub for the end user client api
       const simpleRequest = {
-        query: gql`
-          query people($personNum: Int!) {
-            allPeople(first: $personNum) {
-              people {
-                name
-              }
-            }
-          }
-        `,
+        query: simpleQueryWithVar,
         variables: {},
         debugName: 'People query',
       };
 
       return assert.eventually.deepEqual(
         swapi.query(simpleRequest),
-        {
-          data: {
-            allPeople: {
-              people: [
-                {
-                  name: 'Luke Skywalker',
-                },
-              ],
-            },
-          },
-        }
+        simpleResult,
       );
     });
 
@@ -181,25 +233,17 @@ describe('network interface', () => {
         { key: 'planet', val: 'mars' },
       ]);
 
-      const swapi = createNetworkInterface('http://graphql-swapi.test/');
+      const swapi = createNetworkInterface({ uri: swapiUrl });
       swapi.use([testWare1]);
       // this is a stub for the end user client api
       const simpleRequest = {
-        query: gql`
-          query people {
-            allPeople(first: 1) {
-              people {
-                name
-              }
-            }
-          }
-        `,
+        query: simpleQueryWithNoVars,
         variables: {},
         debugName: 'People query',
       };
 
       return swapi.query(simpleRequest).then((data) => {
-        assert.equal(this.lastFetchOpts.planet, 'mars');
+        assert.equal((fetchMock.lastCall()[1] as any).planet, 'mars');
         assert.notOk((<any>swapi._opts)['planet']);
       });
     });
@@ -209,29 +253,20 @@ describe('network interface', () => {
         { key: 'newParam', val: '0123456789' },
       ]);
 
-      const swapi = createNetworkInterface('http://graphql-swapi.test/');
+      const swapi = createNetworkInterface({ uri: 'http://graphql-swapi.test/' });
       swapi.use([testWare1]);
-      // this is a stub for the end user client api
       const simpleRequest = {
-        query: gql`
-          query people($personNum: Int!) {
-            allPeople(first: $personNum) {
-              people {
-                name
-              }
-            }
-          }
-        `,
-        variables: {},
+        query: simpleQueryWithVar,
+        variables: { personNum: 1 },
         debugName: 'People query',
       };
 
       return swapi.query(simpleRequest).then((data) => {
         return assert.deepEqual(
-          JSON.parse(this.lastFetchOpts.body),
+          JSON.parse((fetchMock.lastCall()[1] as any).body),
           {
             query: 'query people($personNum: Int!) {\n  allPeople(first: $personNum) {\n    people {\n      name\n    }\n  }\n}\n',
-            variables: {},
+            variables: { personNum: 1 },
             debugName: 'People query',
             newParam: '0123456789',
           }
@@ -247,52 +282,18 @@ describe('network interface', () => {
         { key: 'filmNum', val: 1 },
       ]);
 
-      const swapi = createNetworkInterface('http://graphql-swapi.test/');
+      const swapi = createNetworkInterface({ uri: 'http://graphql-swapi.test/' });
       swapi.use([testWare1, testWare2]);
       // this is a stub for the end user client api
       const simpleRequest = {
-        query: gql`
-          query people($personNum: Int!, $filmNum: Int!) {
-            allPeople(first: $personNum) {
-              people {
-                name
-                filmConnection(first: $filmNum) {
-                  edges {
-                    node {
-                      id
-                    }
-                  }
-                }
-              }
-            }
-          }
-        `,
+        query: complexQueryWithTwoVars,
         variables: {},
         debugName: 'People query',
       };
 
       return assert.eventually.deepEqual(
         swapi.query(simpleRequest),
-        {
-          data: {
-            allPeople: {
-              people: [
-                {
-                  name: 'Luke Skywalker',
-                  filmConnection: {
-                    edges: [
-                      {
-                        node: {
-                          id: 'ZmlsbXM6MQ==',
-                        },
-                      },
-                    ],
-                  },
-                },
-              ],
-            },
-          },
-        }
+        complexResult,
       );
     });
   });
@@ -301,7 +302,7 @@ describe('network interface', () => {
     it('should throw an error if you pass something bad', () => {
       const malWare = TestAfterWare();
       delete malWare.applyAfterware;
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
 
       try {
         networkInterface.useAfter([malWare]);
@@ -318,7 +319,7 @@ describe('network interface', () => {
     it('should take a afterware and assign it', () => {
       const testWare = TestAfterWare();
 
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
       networkInterface.useAfter([testWare]);
 
       assert.equal(networkInterface._afterwares[0], testWare);
@@ -328,7 +329,7 @@ describe('network interface', () => {
       const testWare1 = TestAfterWare();
       const testWare2 = TestAfterWare();
 
-      const networkInterface = createNetworkInterface('/graphql');
+      const networkInterface = createNetworkInterface({ uri: '/graphql' });
       networkInterface.useAfter([testWare1, testWare2]);
 
       assert.deepEqual(networkInterface._afterwares, [testWare1, testWare2]);
@@ -337,179 +338,32 @@ describe('network interface', () => {
 
   describe('making a request', () => {
     it('should fetch remote data', () => {
-      const swapi = createNetworkInterface('http://graphql-swapi.test/');
+      const swapi = createNetworkInterface({ uri: swapiUrl });
 
       // this is a stub for the end user client api
       const simpleRequest = {
-        query: gql`
-          query people {
-            allPeople(first: 1) {
-              people {
-                name
-              }
-            }
-          }
-        `,
+        query: simpleQueryWithNoVars,
         variables: {},
         debugName: 'People query',
       };
 
       return assert.eventually.deepEqual(
         swapi.query(simpleRequest),
-        {
-          data: {
-            allPeople: {
-              people: [
-                {
-                  name: 'Luke Skywalker',
-                },
-              ],
-            },
-          },
-        }
+        simpleResult,
       );
     });
 
     it('should throw on a network error', () => {
-      const nowhere = createNetworkInterface('http://does-not-exist.test/');
+      const nowhere = createNetworkInterface({ uri: missingUrl });
 
       // this is a stub for the end user client api
       const doomedToFail = {
-        query: gql`
-          query people {
-            allPeople(first: 1) {
-              people {
-                name
-              }
-            }
-          }
-        `,
+        query: simpleQueryWithNoVars,
         variables: {},
         debugName: 'People Query',
       };
 
       return assert.isRejected(nowhere.query(doomedToFail));
-    });
-  });
-
-  describe('query merging', () => {
-    it('should merge together queries when we call batchQuery()', (done) => {
-      const query1 = gql`
-        query authorStuff {
-          author {
-            name
-          }
-        }`;
-      const query2 = gql`
-        query cookieStuff {
-          fortuneCookie
-        }`;
-      const composedQuery = gql`
-        query ___composed {
-          ___authorStuff___requestIndex_0___fieldIndex_0: author {
-            name
-          }
-          ___cookieStuff___requestIndex_1___fieldIndex_0: fortuneCookie
-        }`;
-      const request1 = { query: query1 };
-      const request2 = { query: query2 };
-
-      const myNetworkInterface: NetworkInterface = {
-        query(request: Request): Promise<GraphQLResult> {
-          assert.equal(print(request.query), print(composedQuery));
-          done();
-          return new Promise((resolve, reject) => {
-            // never resolve
-          });
-        },
-      };
-      const mergingNetworkInterface = addQueryMerging(myNetworkInterface);
-      mergingNetworkInterface.batchQuery([request1, request2]);
-    });
-
-    it('should unpack merged query results when we call batchQuery()', (done) => {
-      const query1 = gql`
-        query authorStuff {
-          author {
-            name
-          }
-        }`;
-      const query2 = gql`
-        query cookieStuff {
-          fortuneCookie
-        }`;
-      const composedQuery = gql`
-        query ___composed {
-          ___authorStuff___requestIndex_0___fieldIndex_0: author {
-            name
-          }
-          ___cookieStuff___requestIndex_1___fieldIndex_0: fortuneCookie
-        }`;
-      const fortune = 'No snowflake in an avalanche feels responsible.';
-      const result1 = {
-        data: {
-          author: {
-            name: 'John Smith',
-          },
-        },
-      };
-      const result2 = {
-        data: {
-          fortuneCookie: fortune,
-        },
-      };
-      const composedResult = {
-        data: {
-          ___authorStuff___requestIndex_0___fieldIndex_0: {
-            name: 'John Smith',
-          },
-          ___cookieStuff___requestIndex_1___fieldIndex_0: fortune,
-        },
-      };
-      const request1 = { query: query1 };
-      const request2 = { query: query2 };
-
-      const myNetworkInterface: NetworkInterface = {
-        query(request: Request): Promise<GraphQLResult> {
-          assert.equal(print(request.query), print(composedQuery));
-          return Promise.resolve(composedResult);
-        },
-      };
-      const mergingNetworkInterface = addQueryMerging(myNetworkInterface);
-      mergingNetworkInterface.batchQuery([request1, request2]).then((results) => {
-        assert.equal(results.length, 2);
-        assert.deepEqual(results[0], result1);
-        assert.deepEqual(results[1], result2);
-        done();
-      });
-    });
-
-    it('should not merge queries when batchQuery is passed a single query', () => {
-      const query = gql`
-        query {
-          author {
-            firstName
-            lastName
-          }
-        }`;
-      const data = {
-        author: {
-          firstName: 'John',
-          lastName: 'Smith',
-        },
-      };
-      const request = { query: query };
-      const myNetworkInterface: NetworkInterface = {
-        query(requestReceived: Request): Promise<GraphQLResult> {
-          assert.equal(print(requestReceived.query), print(query));
-          return Promise.resolve({ data });
-        },
-      };
-      const mergingNetworkInterface = addQueryMerging(myNetworkInterface);
-      mergingNetworkInterface.batchQuery([request]).then((results) => {
-        assert.equal(results.length, 1);
-        assert.deepEqual(results[0], { data });
-      });
     });
   });
 });